剥离 http 请求.
