+++ lib/sensors_analytics_sdk.rb
@@ -1,6 +1,7 @@
 
 require 'sensors_analytics/version'
 require 'sensors_analytics/errors'
+require 'sensors_analytics/http'
 require 'sensors_analytics/consumers'
 require 'sensors_analytics/client'
